+++ module/browse.php
@@ -111,7 +111,9 @@ $labelKeys = array(
 	'DoliCatalogEmptyCategory', 'DoliCatalogError', 'DoliCatalogFavorites',
 	'DoliCatalogRecent', 'DoliCatalogOpenProduct', 'DoliCatalogAddFavorite',
 	'DoliCatalogRemoveFavorite', 'DoliCatalogPrevious', 'DoliCatalogNext',
-	'DoliCatalogBrowseEmpty', 'DoliCatalogRefineBy', 'DoliCatalogClearTags',
+	'DoliCatalogBrowseEmpty', 'DoliCatalogRefineBy', 'DoliCatalogClearTags', 'DoliCatalogOtherTags', 'DoliCatalogMoreValues',
+	'DoliCatalogMatchAll', 'DoliCatalogMatchAny', 'DoliCatalogMatchAllHint', 'DoliCatalogMatchAnyHint',
+	'DoliCatalogFiltersActive',
 );
 foreach ($labelKeys as $k) {
 	$config['labels'][$k] = $langs->transnoentities($k);